Extract the cache functions from the GitHubRepository to its own class
